refactor: consolidate npm exports and trace flock acquisition
